requirements.txt
pandas
pandas_ta
sqlalchemy
numpy
datetime
yfinance
discord.py==2.3.2
bs4
matplotlib
psycopg2-binary
ta
gspread
gspread_dataframe